wip: move google credential manager to app module
to provide MainActiviry as context
