+++ feat/auth/build.gradle.kts
@@ -1,8 +1,3 @@
-import com.android.build.gradle.internal.cxx.configure.gradleLocalProperties
-
-fun getLocalProperty(propertyKey: String) =
-    System.getenv(propertyKey) ?: gradleLocalProperties(rootDir, providers).getProperty(propertyKey)
-
 plugins {
     id("com.emotionstorage.convention.android.library")
     id("com.emotionstorage.convention.android.library.compose")
@@ -15,12 +10,6 @@ android {
     defaultConfig {
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
 //        consumerProguardFiles("consumer-rules.pro")
-
-        buildConfigField(
-            "String",
-            "GOOGLE_SERVER_CLIENT_ID",
-            getLocalProperty("GOOGLE_SERVER_CLIENT_ID"),
-        )
     }
 
     buildTypes {
@@ -43,6 +32,4 @@ dependencies {
     implementation(projects.core.common)
     implementation(projects.core.presentation)
     implementation(projects.core.ui)
-
-    implementation(libs.bundles.credentials)
 }
